requirements.txt
flask
gensim
numpy
pandas
google-generativeai
scikit-learn
spacy